ChatAPI/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.0.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
msgpack==1.0.4
PyJWT==1.7.1
pytz==2022.7.1
redis==4.4.2
sqlparse==0.4.3
